cfg: introduce grammar-object map with lazily evaluated values
...as a possible solution to share data, such as follow set or
state-items-mapping of a grammar, that requires expensive computation
to generate, between functions.

Previous approach of simply passing such data through function
parameters has two problems:
 1) parameters that depend on one another, e.g. g and
    state-items-mappings, which can be generated from g;
 2) long parameter-list, within which some parameters' purpose is
    unclear to user of the functions.

While eliminating the generated parameters, such as state-items-mapping,
resolves the two problems listed above, it can introduce
 3) heavy performance penalty by doing expensive computation within
    a function to generate necessary intermediate data every time
    the function is called.

On the other hand, by generating related functions out of a
closure, where expensive data is generated once and for all, all
of problems 1), 2), and 3) are resolved. However, it
 4) introduces much less readable code;
 5) makes writing test for the related functions difficult;
 6) makes using the related functions in repl cumbersome.

Thus, the proposed solution in this commit @~@